Move GUI Handles for Folders, Tabs, and Markdown to _gui_handles.py
